+++ .changeset/three-stingrays-switch.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/typography': patch
+---
+
+Refactors component to follow new directory file structure
