+++ packages/typography/src/Description/Description.types.ts
@@ -1,0 +1,6 @@
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+
+export type DescriptionProps = HTMLElementProps<'p', never> & {
+    darkMode?: boolean;
+    disabled?: boolean;
+};
