+++ packages/typography/src/H1/H1.types.ts
@@ -1,0 +1,4 @@
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+import { CommonTypographyProps } from '../types';
+
+export type H1Props = HTMLElementProps<'h1'> & CommonTypographyProps;
