+++ packages/typography/src/H2/H2.types.ts
@@ -1,0 +1,4 @@
+import { CommonTypographyProps } from '../types';
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+
+export type H2Props = HTMLElementProps<'h2'> & CommonTypographyProps;
