+++ packages/typography/src/H3/H3.styles.ts
@@ -1,0 +1,7 @@
+import { css } from '@leafygreen-ui/emotion';
+
+export const h3Styles = css`
+  font-size: 24px;
+  line-height: 32px;
+  font-weight: 500;
+`;
