+++ packages/typography/src/H3/H3.tsx
@@ -1,21 +1,11 @@
 import React from 'react';
 import Box, { ExtendableBox } from '@leafygreen-ui/box';
-import { css, cx } from '@leafygreen-ui/emotion';
-import { HTMLElementProps } from '@leafygreen-ui/lib';
-import { baseTypographyStyles, defaultTextColor } from './styles';
-import { CommonTypographyProps } from './types';
+import { cx } from '@leafygreen-ui/emotion';
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
+import { baseTypographyStyles, defaultTextColor } from '../styles';
+import { h3Styles } from './H3.styles'
+import { H3Props } from './H3.types'
 
-/**
- * H3
- */
-const h3 = css`
-  font-size: 24px;
-  line-height: 32px;
-  font-weight: 500;
-`;
-
-type H3Props = HTMLElementProps<'h3'> & CommonTypographyProps;
 
 const H3: ExtendableBox<H3Props, 'h3'> = ({
   darkMode: darkModeProp,
@@ -28,7 +18,7 @@ const H3: ExtendableBox<H3Props, 'h3'> = ({
       as="h3"
       className={cx(
         baseTypographyStyles,
-        h3,
+        h3Styles,
         defaultTextColor[theme],
         className,
       )}
