+++ packages/typography/src/InlineCode/InlineCode.types.ts
@@ -1,0 +1,8 @@
+import { HTMLElementProps, OneOf } from '@leafygreen-ui/lib';
+import { CommonTypographyProps } from '../types';
+
+export type InlineCodeProps = OneOf<
+    HTMLElementProps<'code'>,
+    HTMLElementProps<'a'>
+> &
+    CommonTypographyProps;
