+++ packages/typography/src/InlineKeyCode/InlineKeyCode.types.ts
@@ -1,0 +1,4 @@
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+import { CommonTypographyProps } from '../types';
+
+export type InlineKeyCodeProps = HTMLElementProps<'h1'> & CommonTypographyProps;
