+++ packages/typography/src/Label/Label.types.ts
@@ -1,0 +1,7 @@
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+
+export type LabelProps = HTMLElementProps<'label', never> & {
+    darkMode?: boolean;
+    htmlFor: string;
+    disabled?: boolean;
+};
