+++ packages/typography/src/Overline/Overline.tsx
@@ -1,23 +1,10 @@
 import React from 'react';
 import Box, { ExtendableBox } from '@leafygreen-ui/box';
-import { css, cx } from '@leafygreen-ui/emotion';
-import { HTMLElementProps } from '@leafygreen-ui/lib';
-import { baseTypographyStyles, defaultTextColor } from './styles';
-import { CommonTypographyProps } from './types';
+import { cx } from '@leafygreen-ui/emotion';
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
-
-/**
- * Overline
- */
-const overline = css`
-  font-size: 12px;
-  font-weight: 700;
-  text-transform: uppercase;
-  line-height: 20px;
-  letter-spacing: 0.4px;
-`;
-
-type OverlineProps = HTMLElementProps<'div'> & CommonTypographyProps;
+import { baseTypographyStyles, defaultTextColor } from '../styles';
+import { overlineStyles } from './Overline.styles'
+import { OverlineProps } from './Overline.types';
 
 export const Overline: ExtendableBox<OverlineProps, 'div'> = ({
   darkMode: darkModeProp,
@@ -29,7 +16,7 @@ export const Overline: ExtendableBox<OverlineProps, 'div'> = ({
     <Box
       className={cx(
         baseTypographyStyles,
-        overline,
+        overlineStyles,
         defaultTextColor[theme],
         className,
       )}
