+++ packages/typography/src/Overline/Overline.types.ts
@@ -1,0 +1,4 @@
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+import { CommonTypographyProps } from '../types';
+
+export type OverlineProps = HTMLElementProps<'div'> & CommonTypographyProps;
