+++ packages/typography/src/Subtitle/Subtitle.styles.ts
@@ -1,0 +1,7 @@
+import { css } from '@leafygreen-ui/emotion'
+
+export const subtitleStyles = css`
+  font-size: 18px;
+  line-height: 24px;
+  font-weight: 700;
+`;
