+++ packages/typography/src/Subtitle/Subtitle.tsx
@@ -1,21 +1,10 @@
 import React from 'react';
 import Box, { ExtendableBox } from '@leafygreen-ui/box';
-import { css, cx } from '@leafygreen-ui/emotion';
-import { HTMLElementProps } from '@leafygreen-ui/lib';
-import { baseTypographyStyles, defaultTextColor } from './styles';
-import { CommonTypographyProps } from './types';
+import { cx } from '@leafygreen-ui/emotion';
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
-
-/**
- * Subtitle
- */
-const subtitle = css`
-  font-size: 18px;
-  line-height: 24px;
-  font-weight: 700;
-`;
-
-type SubtitleProps = HTMLElementProps<'h6'> & CommonTypographyProps;
+import { baseTypographyStyles, defaultTextColor } from '../styles';
+import { subtitleStyles } from './Subtitle.styles'
+import { SubtitleProps } from './Subtitle.types';
 
 const Subtitle: ExtendableBox<SubtitleProps, 'h6'> = ({
   darkMode: darkModeProp,
@@ -28,7 +17,7 @@ const Subtitle: ExtendableBox<SubtitleProps, 'h6'> = ({
       as="h6"
       className={cx(
         baseTypographyStyles,
-        subtitle,
+        subtitleStyles,
         defaultTextColor[theme],
         className,
       )}
