+++ packages/typography/src/Subtitle/Subtitle.types.ts
@@ -1,0 +1,4 @@
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+import { CommonTypographyProps } from '../types';
+
+export type SubtitleProps = HTMLElementProps<'h6'> & CommonTypographyProps;
