+++ packages/typography/src/index.ts
@@ -1,15 +1,17 @@
-export { useUpdatedBaseFontSize } from './useUpdatedBaseFontSize';
-export { default as Body } from './Body';
-export { default as Description } from './Description';
-export { default as Disclaimer } from './Disclaimer';
-export { default as H1 } from './H1';
-export { default as H2 } from './H2';
-export { default as H3 } from './H3';
-export { default as InlineCode } from './InlineCode';
-export { default as InlineKeyCode } from './InlineKeyCode';
-export { default as Label } from './Label';
-export { default as Link, anchorClassName } from './Link';
-export { default as Overline } from './Overline';
-export { default as Subtitle } from './Subtitle';
-export { bodyTypeScaleStyles, labelTypeScaleStyles } from './styles';
-export { StaticWidthText } from './utility/StaticWidthText';
+export { default as Body } from './Body/Body';
+export { default as Description } from './Description/Description';
+export { default as Disclaimer } from './Disclaimer/Disclaimer';
+export { default as H1 } from './H1/H1';
+export { default as H2 } from './H2/H2';
+export { default as H3 } from './H3/H3';
+export { default as InlineCode } from './InlineCode/InlineCode';
+export { default as InlineKeyCode } from './InlineKeyCode/InlineKeyCode';
+export { default as Label } from './Label/Label';
+export { labelTypeScaleStyles } from './Label/Label.styles'
+export { default as Link } from './Link/Link';
+export { anchorClassName } from './Link/Link.styles'
+export { default as Overline } from './Overline/Overline';
+export { default as Subtitle } from './Subtitle/Subtitle';
+export { bodyTypeScaleStyles } from './styles';
+export { StaticWidthText } from './utils/StaticWidthText';
+export { useUpdatedBaseFontSize } from './utils/useUpdatedBaseFontSize';
