+++ packages/typography/src/utils/StaticWidthText.tsx
