+++ packages/typography/tsconfig.json
@@ -6,7 +6,8 @@
     "rootDir": "src"
   },
   "include": [
-    "src/**/*"
+    "src/**/*",
+    "H1/H1.tsx"
   ],
   "references": [
     {
